修复优化
修复优化

+++ src/main/java/com/shengchanshe/chang_sheng_jue/block/custom/brick_kiln/BrickKiln.java
@@ -95,7 +95,7 @@ public class BrickKiln extends BaseEntityBlock {
 
     @Override
     public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pIsMoving) {
-        if(pState.getBlock() != pNewState.getBlock()){
+        if(!pLevel.isClientSide && pState.getBlock() != pNewState.getBlock()){
             BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
             if(blockEntity instanceof BrickKilnEntity){
                 ((BrickKilnEntity) blockEntity).drop();
