+++ src/main/java/com/shengchanshe/chang_sheng_jue/block/custom/forgeblock/ForgeBlock.java
@@ -49,7 +49,7 @@ public class ForgeBlock extends BaseEntityBlock {
 
     @Override
     public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pIsMoving) {
-        if(pState.getBlock() != pNewState.getBlock()){
+        if(!pLevel.isClientSide && pState.getBlock() != pNewState.getBlock()){
             BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
             if(blockEntity instanceof ForgeBlockEntity){
                 ((ForgeBlockEntity) blockEntity).drop();
