+++ src/main/java/com/shengchanshe/chang_sheng_jue/block/custom/tailoringcase/TailoringCase.java
@@ -48,7 +48,7 @@ public class TailoringCase extends BaseEntityBlock {
 
     @Override
     public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pIsMoving) {
-        if(pState.getBlock() != pNewState.getBlock()){
+        if(!pLevel.isClientSide && pState.getBlock() != pNewState.getBlock()){
             BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
             if(blockEntity instanceof TailoringCaseEntity){
                 ((TailoringCaseEntity) blockEntity).drop();
