+++ src/main/java/com/shengchanshe/chang_sheng_jue/block/custom/workbench/WoodworkingBench.java
@@ -41,7 +41,7 @@ public class WoodworkingBench extends BaseEntityBlock {
 
     @Override
     public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pIsMoving) {
-        if(pState.getBlock() != pNewState.getBlock()){
+        if(!pLevel.isClientSide && pState.getBlock() != pNewState.getBlock()){
             BlockEntity blockEntity = pLevel.getBlockEntity(pPos);
             if(blockEntity instanceof WoodworkingBenchEntity){
                 ((WoodworkingBenchEntity) blockEntity).drop();
